Organize code files
