+++ packages/shared-ux/chrome/navigation/src/i18n_strings.ts
