+++ packages/shared-ux/chrome/navigation/src/ui/components/elastic_mark.tsx
