+++ packages/shared-ux/chrome/navigation/src/ui/components/header_logo.scss
