+++ packages/shared-ux/chrome/navigation/src/ui/components/index.tsx
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+export { NavHeader } from './header';
+export { LinkToCloud } from './link_to_cloud';
+export { NavigationBucket } from './navigation_bucket';
+export { RecentlyAccessed } from './recently_accessed';
